✅ Splitter opp tester for saksbehandlingsperiode i flere filer
Nye tester: DagoversiktLogikkTest, DokumentHentingTest, InntektsforholdOperasjonerTest,
SaksbehandlingsperiodeOpprettelseTest, SaksbehandlingsperiodeTestFixtures,
SaksbehandlingsperiodeValideringTest.
Fjernet: SaksbehandlingFlytTest.
